backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
redis==7.4.0
celery==5.4.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
transformers==4.44.2
torch==2.4.1
openai-whisper==20231117
librosa==0.10.2.post1
numpy==1.26.4
scikit-learn==1.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
